Renaming _default_experiment folder to default_experiment
