+++ tests/resources/mlflow-test-plugin/mlflow_test_plugin/default_experiment_provider.py
@@ -1,4 +1,4 @@
-from mlflow.tracking._default_experiment.abstract_context import DefaultExperimentProvider
+from mlflow.tracking.default_experiment.abstract_context import DefaultExperimentProvider
 
 
 class PluginDefaultExperimentProvider(DefaultExperimentProvider):
